+++ inc/BuildAssets.php
@@ -27,7 +27,7 @@ class BuildAssets {
      *
      * @var array
      */
-    private static $appCssDependencies = array( 'wp-components', 'bluehost-font' );
+    private static $appCssDependencies = array( 'wpadmin-brand-bluehost', 'wp-components' );
 
     /**
      * Run asset registration globally. Use self::enqueue() in scoped contexts.
@@ -71,8 +71,8 @@ class BuildAssets {
 		);
 
 		wp_register_style(
-			'bluehost-font',
-			'https://fonts.googleapis.com/css?family=Open+Sans:300,400,600',
+			'wpadmin-brand-bluehost',
+            $url . 'static/wpadmin-brand-bluehost.css',
 			array(),
 			empty( $min ) ? $rand : BLUEHOST_PLUGIN_VERSION
 		);
