+++ src/app/app-main.scss
@@ -27,6 +27,7 @@
  }
 
  #bwa-app {
+   background-color: #f1f5f7;
    will-change: opacity;
    @extend .is-font;
    @extend .appFadeIn;
