+++ src/app/components/atoms/bwa-box/index.js
@@ -1,0 +1,11 @@
+import './style.scss'
+
+export const BWABox = ({children}) => {
+    return (
+        <div className="bwa-box">
+            {children}
+        </div>
+    )
+}
+
+export default BWABox;
