+++ src/app/components/atoms/bwa-box/style.scss
@@ -1,0 +1,8 @@
+.bwa-box {
+    background: #fff;
+    border: 1px solid #ddd;
+    border-radius: 3px;
+    padding: 1rem;
+    font-size: 16px;
+    line-height: 24px;
+}
