+++ src/app/pages/home/onboarding/steps/LEGACYV1-homepage/index.js
