+++ src/app/pages/home/onboarding/steps/base.js
@@ -1,0 +1,7 @@
+export const BaseStep = ({ children }) => (
+    <div className="nf-onboarding-base-step">
+        {children}
+    </div>
+);
+
+export default BaseStep;
