+++ src/app/pages/home/onboarding/style.scss
@@ -1,137 +1,202 @@
 @import '@app/config.scss';
 
-.bluehost-onboarding {
-
-  line-height: 24px;
-
-  a {
-    color: #2C6CC9;
-  }
-
-  &-page-header {
-    &__title {
-      font-size: 32px;
-      font-weight: 300;
-      line-height: 40px;
-    }
-
-    &__description {
-      margin-bottom: 30px;
-    }
-  }
-
-  &-page-content {
-
-  }
-
-  &-page-footer {
-
-    padding-top: 22px;
-
-    &__title {
-      font-size: 20px;
-      font-weight: 600;
-      line-height: 32px;
-    }
-
-    &__content {
-
-    }
-  }
-
-  &-steps {
-
-  }
-
-  &-step {
-    background-color: $color-white;
-    border-radius: 0 3px 3px 0;
-    box-shadow: 0 2px 5px 0 rgba(42, 47, 69, 0.08), 0 1px 1.5px 0 rgba(0, 0, 0, 0.08), 0 1px 2px 0 rgba(0, 0, 0, 0.08);
-    margin: 0 0 18px 55px;
-    padding: 20px 20px 20px 75px;
-    position: relative;
-
-    &__image {
-      position: absolute;
-      width: 110px;
-      height: 110px;
-      top: calc(50% - 55px);
-      left: -55px;
-    }
-
-    &__title {
-      font-size: 20px;
-      font-weight: 600;
-      line-height: 32px;
-    }
-
-    &__description {
-      max-width: 500px;
-      margin: 0 0 1rem 0;
-    }
-
-    &__actions {
-      display: flex;
-      justify-content: space-between;
-
-      a:only-child,
-      button:only-child {
-        margin-left: auto;
-      }
-
-      .components-button.is-link {
-        margin: 0;
-        padding-left: 0 !important;
-        font-weight: 600;
-      }
-    }
-
-    &.--launch {
-      border-left: 4px solid $color-orange;
-    }
-
-    &.--launched {
-      border-left: 4px solid $color-green;
-    }
-
-    &.--homepage-complete {
-      border-left: 4px solid $color-green;
-    }
-
-    &.--homepage-open {
-      .bluehost-onboarding-step__content {
-        align-items: center;
-        display: flex;
-        flex-direction: column;
-        margin: 2rem auto 0 auto;
-        width: max-content;
-      }
-
-      .icon-collection {
-        align-self: stretch;
-        align-items: center;
-        display: flex;
-        justify-content: space-between;
-        margin-bottom: 30px;
-      }
-
-      .icon-block {
-        background: none;
-        border: none;
-        outline: none;
-
-        p {
-          color: #3575D3;
-          font-size: 16px;
-          font-weight: 600;
-          letter-spacing: 0;
-          line-height: 24px;
-          text-align: center;
-          margin: 0;
-        }
-      }
-
-    }
-  }
-
+.grid-col-2 {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  margin-top: 2rem;
+  grid-column-gap: 2rem;
 }
+
+.bluehost-onboarding-page-header {
+  .onboarding-close {
+    float: right;
+    position: relative;
+    top: 2rem;
+    right: 1rem;
+    border: 0;
+    background: transparent;
+  }
+}
+
+.nf-onboarding-tabs {
+  display: flex;
+  flex-direction: row;
+  .components-tab-panel {
+    &__tabs {
+      width: 240px;
+      min-width: 240px;
+    }
+    &__tabs-item {
+      text-align: left;
+      // border-radius: 3px;
+      border-left: 4px solid transparent;
+      font-size: 1rem;
+      &.is-active {
+        border-left-color: var(--nf-primary);
+        color: var(--nf-primary);
+        box-shadow: none;
+        font-weight: 500;
+      }
+      &:focus:not(:disabled),
+      &:hover:not(:disabled) {
+        // box-shadow: inset 0 0 0 var(--wp-admin-border-width-focus) var(--wp-admin-theme-color);
+        box-shadow: none;
+        color: var(--nf-primary);
+      }
+    }
+    &__tab-content {
+      padding-left: 1rem;
+    }
+  }
+}
+
+.nf-onboarding-base-step {
+  max-width: 735px;
+  img {
+    display: block;
+    height: auto;
+    margin: 0 auto;
+    max-width: 425px !important;
+    max-height: 425px !important;
+  }
+  p {
+    font-size: 1.25rem;
+    line-height: 1.5;
+  }
+}
+
+// .bluehost-onboarding {
+
+//   line-height: 24px;
+
+//   a {
+//     color: #2C6CC9;
+//   }
+
+//   &-page-header {
+//     &__title {
+//       font-size: 32px;
+//       font-weight: 300;
+//       line-height: 40px;
+//     }
+
+//     &__description {
+//       margin-bottom: 30px;
+//     }
+//   }
+
+//   &-page-content {
+
+//   }
+
+//   &-page-footer {
+
+//     padding-top: 22px;
+
+//     &__title {
+//       font-size: 20px;
+//       font-weight: 600;
+//       line-height: 32px;
+//     }
+
+//     &__content {
+
+//     }
+//   }
+
+//   &-steps {
+
+//   }
+
+//   &-step {
+//     background-color: $color-white;
+//     border-radius: 0 3px 3px 0;
+//     box-shadow: 0 2px 5px 0 rgba(42, 47, 69, 0.08), 0 1px 1.5px 0 rgba(0, 0, 0, 0.08), 0 1px 2px 0 rgba(0, 0, 0, 0.08);
+//     margin: 0 0 18px 55px;
+//     padding: 20px 20px 20px 75px;
+//     position: relative;
+
+//     &__image {
+//       position: absolute;
+//       width: 110px;
+//       height: 110px;
+//       top: calc(50% - 55px);
+//       left: -55px;
+//     }
+
+//     &__title {
+//       font-size: 20px;
+//       font-weight: 600;
+//       line-height: 32px;
+//     }
+
+//     &__description {
+//       max-width: 500px;
+//       margin: 0 0 1rem 0;
+//     }
+
+//     &__actions {
+//       display: flex;
+//       justify-content: space-between;
+
+//       a:only-child,
+//       button:only-child {
+//         margin-left: auto;
+//       }
+
+//       .components-button.is-link {
+//         margin: 0;
+//         padding-left: 0 !important;
+//         font-weight: 600;
+//       }
+//     }
+
+//     &.--launch {
+//       border-left: 4px solid $color-orange;
+//     }
+
+//     &.--launched {
+//       border-left: 4px solid $color-green;
+//     }
+
+//     &.--homepage-complete {
+//       border-left: 4px solid $color-green;
+//     }
+
+//     &.--homepage-open {
+//       .bluehost-onboarding-step__content {
+//         align-items: center;
+//         display: flex;
+//         flex-direction: column;
+//         margin: 2rem auto 0 auto;
+//         width: max-content;
+//       }
+
+//       .icon-collection {
+//         align-self: stretch;
+//         align-items: center;
+//         display: flex;
+//         justify-content: space-between;
+//         margin-bottom: 30px;
+//       }
+
+//       .icon-block {
+//         background: none;
+//         border: none;
+//         outline: none;
+
+//         p {
+//           color: #3575D3;
+//           font-size: 16px;
+//           font-weight: 600;
+//           letter-spacing: 0;
+//           line-height: 24px;
+//           text-align: center;
+//           margin: 0;
+//         }
+//       }
+
+//     }
+//   }
+
+// }
