--- src/app/pages/home/welcome/style.scss
@@ -1,7 +1,0 @@
-.welcome-section {
-    max-width: 600px;
-    margin-bottom: 48px;
-    p {
-        font-size: 16px;
-    }
-}
